Introduce "mapString" functionality from ts-string-visitor
