+++ package.json
@@ -58,7 +58,7 @@
         "ts-jest": "22.4.6",
         "tslint": "5.10.0",
         "tslint-config-prettier": "1.13.0",
-        "typescript": "2.9.1"
+        "typescript": "3.0.3"
     },
     "peerDependencies": {
         "typescript": ">= 2.8.1"
