+++ tslint.json
@@ -1,8 +1,5 @@
 {
-    "extends": [
-        "tslint:recommended",
-        "tslint-config-prettier"
-    ],
+    "extends": ["tslint:recommended", "tslint-config-prettier"],
     "rules": {
         "typedef": [
             true,
@@ -30,6 +27,7 @@
         "no-namespace": false,
         "no-default-export": true,
         "no-unused-variable": true,
+        "object-literal-sort-keys": false,
         "object-literal-shorthand": false,
         "object-literal-key-quotes": false,
         "max-classes-per-file": [false],
