+++ type_tests/v2_8_plus/EnumWrapper/enum-both.dtslint.ts
@@ -1,9 +1,9 @@
-import { $enum } from "../../dist/types";
+import { $enum } from "../../../dist/types";
 
-// Enum with string values
+// Enum with mix of number and string values
 enum TestEnum {
-    A = "a",
-    B = "b",
+    A,
+    B,
     C = "c"
 }
 
@@ -15,6 +15,10 @@ declare const num: number;
 declare const numOrNull: number | null;
 declare const numOrUndefined: number | undefined;
 
+declare const numstr: number | string;
+declare const numstrOrNull: number | string | null;
+declare const numstrOrUndefined: number | string | undefined;
+
 declare const key: keyof typeof TestEnum;
 declare const keyOrNull: keyof typeof TestEnum | null;
 declare const keyOrUndefined: keyof typeof TestEnum | undefined;
@@ -25,7 +29,7 @@ declare const valueOrUndefined: TestEnum | undefined;
 
 const enumWrapper = $enum(TestEnum);
 
-// $ExpectType EnumWrapper<string, typeof TestEnum>
+// $ExpectType EnumWrapper<string | number, typeof TestEnum>
 enumWrapper;
 
 // $ExpectType number
@@ -68,7 +72,7 @@ enumWrapper.forEach((value, key, collection, index) => {
     value;
     // $ExpectType "A" | "B" | "C"
     key;
-    // $ExpectType EnumWrapper<string, typeof TestEnum>
+    // $ExpectType EnumWrapper<string | number, typeof TestEnum>
     collection;
     // $ExpectType number
     index;
@@ -82,7 +86,7 @@ enumWrapper.map((value, key, collection, index) => {
     value;
     // $ExpectType "A" | "B" | "C"
     key;
-    // $ExpectType EnumWrapper<string, typeof TestEnum>
+    // $ExpectType EnumWrapper<string | number, typeof TestEnum>
     collection;
     // $ExpectType number
     index;
@@ -151,86 +155,78 @@ enumWrapper.asKeyOrDefault(str, str);
 enumWrapper.asKeyOrDefault(str, strOrUndefined);
 
 // $ExpectType boolean
-enumWrapper.isValue(str);
+enumWrapper.isValue(numstr);
 // $ExpectType boolean
-enumWrapper.isValue(strOrNull);
+enumWrapper.isValue(numstrOrNull);
 // $ExpectType boolean
-enumWrapper.isValue(strOrUndefined);
-// $ExpectError
-enumWrapper.isValue(num);
+enumWrapper.isValue(numstrOrUndefined);
 
-if (enumWrapper.isValue(str)) {
+if (enumWrapper.isValue(numstr)) {
     // $ExpectType TestEnum
-    str;
+    numstr;
 }
 
-if (enumWrapper.isValue(strOrNull)) {
+if (enumWrapper.isValue(numstrOrNull)) {
     // $ExpectType TestEnum
-    strOrNull;
+    numstrOrNull;
 }
 
-if (enumWrapper.isValue(strOrUndefined)) {
+if (enumWrapper.isValue(numstrOrUndefined)) {
     // $ExpectType TestEnum
-    strOrUndefined;
+    numstrOrUndefined;
 }
 
 // $ExpectType TestEnum
-enumWrapper.asValueOrThrow(str);
+enumWrapper.asValueOrThrow(numstr);
 // $ExpectType TestEnum
-enumWrapper.asValueOrThrow(strOrNull);
+enumWrapper.asValueOrThrow(numstrOrNull);
 // $ExpectType TestEnum
-enumWrapper.asValueOrThrow(strOrUndefined);
-// $ExpectError
-enumWrapper.asValueOrThrow(num);
+enumWrapper.asValueOrThrow(numstrOrUndefined);
 
 // $ExpectType TestEnum | undefined
-enumWrapper.asValueOrDefault(str);
+enumWrapper.asValueOrDefault(numstr);
 // $ExpectType TestEnum | undefined
-enumWrapper.asValueOrDefault(strOrNull);
+enumWrapper.asValueOrDefault(numstrOrNull);
 // $ExpectType TestEnum | undefined
-enumWrapper.asValueOrDefault(strOrUndefined);
-// $ExpectError
-enumWrapper.asValueOrDefault(num);
+enumWrapper.asValueOrDefault(numstrOrUndefined);
 
 // $ExpectType TestEnum | undefined
-enumWrapper.asValueOrDefault(str, undefined);
+enumWrapper.asValueOrDefault(numstr, undefined);
 // $ExpectType TestEnum
-enumWrapper.asValueOrDefault(str, value);
+enumWrapper.asValueOrDefault(numstr, value);
 // $ExpectType TestEnum | undefined
-enumWrapper.asValueOrDefault(str, valueOrUndefined);
+enumWrapper.asValueOrDefault(numstr, valueOrUndefined);
+// $ExpectType string | number
+enumWrapper.asValueOrDefault(numstr, num);
+// $ExpectType string | number
+enumWrapper.asValueOrDefault(numstr, str);
+// $ExpectType string | number
+enumWrapper.asValueOrDefault(numstr, numstr);
+// $ExpectType string | number | undefined
+enumWrapper.asValueOrDefault(num, numstrOrUndefined);
+
+// $ExpectType "A" | "B" | "C"
+enumWrapper.getKeyOrThrow(numstr);
+// $ExpectType "A" | "B" | "C"
+enumWrapper.getKeyOrThrow(numstrOrNull);
+// $ExpectType "A" | "B" | "C"
+enumWrapper.getKeyOrThrow(numstrOrUndefined);
+
+// $ExpectType "A" | "B" | "C" | undefined
+enumWrapper.getKeyOrDefault(numstr);
+// $ExpectType "A" | "B" | "C" | undefined
+enumWrapper.getKeyOrDefault(numstrOrNull);
+// $ExpectType "A" | "B" | "C" | undefined
+enumWrapper.getKeyOrDefault(numstrOrUndefined);
+
+// $ExpectType "A" | "B" | "C"
+enumWrapper.getKeyOrDefault(numstr, key);
+// $ExpectType "A" | "B" | "C" | undefined
+enumWrapper.getKeyOrDefault(numstr, keyOrUndefined);
 // $ExpectType string
-enumWrapper.asValueOrDefault(str, str);
+enumWrapper.getKeyOrDefault(numstr, str);
 // $ExpectType string | undefined
-enumWrapper.asValueOrDefault(str, strOrUndefined);
-// $ExpectError
-enumWrapper.asValueOrDefault(str, num);
-
-// $ExpectType "A" | "B" | "C"
-enumWrapper.getKeyOrThrow(str);
-// $ExpectType "A" | "B" | "C"
-enumWrapper.getKeyOrThrow(strOrNull);
-// $ExpectType "A" | "B" | "C"
-enumWrapper.getKeyOrThrow(strOrUndefined);
-// $ExpectError
-enumWrapper.getKeyOrThrow(num);
-
-// $ExpectType "A" | "B" | "C" | undefined
-enumWrapper.getKeyOrDefault(str);
-// $ExpectType "A" | "B" | "C" | undefined
-enumWrapper.getKeyOrDefault(strOrNull);
-// $ExpectType "A" | "B" | "C" | undefined
-enumWrapper.getKeyOrDefault(strOrUndefined);
-// $ExpectError
-enumWrapper.getKeyOrDefault(num);
-
-// $ExpectType "A" | "B" | "C"
-enumWrapper.getKeyOrDefault(str, key);
-// $ExpectType "A" | "B" | "C" | undefined
-enumWrapper.getKeyOrDefault(str, keyOrUndefined);
-// $ExpectType string
-enumWrapper.getKeyOrDefault(str, str);
-// $ExpectType string | undefined
-enumWrapper.getKeyOrDefault(str, strOrUndefined);
+enumWrapper.getKeyOrDefault(numstr, strOrUndefined);
 
 // $ExpectType TestEnum
 enumWrapper.getValueOrThrow(key);
@@ -258,9 +254,11 @@ enumWrapper.getValueOrDefault(str, undefined);
 enumWrapper.getValueOrDefault(str, value);
 // $ExpectType TestEnum | undefined
 enumWrapper.getValueOrDefault(str, valueOrUndefined);
-// $ExpectType string
-enumWrapper.getValueOrDefault(str, str);
-// $ExpectType string | undefined
-enumWrapper.getValueOrDefault(str, strOrUndefined);
-// $ExpectError
+// $ExpectType string | number
 enumWrapper.getValueOrDefault(str, num);
+// $ExpectType string | number
+enumWrapper.getValueOrDefault(str, str);
+// $ExpectType string | number
+enumWrapper.getValueOrDefault(str, numstr);
+// $ExpectType string | number | undefined
+enumWrapper.getValueOrDefault(str, numstrOrUndefined);
