+++ type_tests/v2_8_plus/EnumWrapper/enum-string.dtslint.ts
@@ -1,11 +1,11 @@
-import { $enum } from "../../dist/types";
+import { $enum } from "../../../dist/types";
 
-// Enum-like object with mix of number and string values
-const TestEnum = {
-    A: 1,
-    B: 2,
-    C: "c"
-};
+// Enum with string values
+enum TestEnum {
+    A = "a",
+    B = "b",
+    C = "c"
+}
 
 declare const str: string;
 declare const strOrNull: string | null;
@@ -15,17 +15,17 @@ declare const num: number;
 declare const numOrNull: number | null;
 declare const numOrUndefined: number | undefined;
 
-declare const numstr: number | string;
-declare const numstrOrNull: number | string | null;
-declare const numstrOrUndefined: number | string | undefined;
-
 declare const key: keyof typeof TestEnum;
 declare const keyOrNull: keyof typeof TestEnum | null;
 declare const keyOrUndefined: keyof typeof TestEnum | undefined;
 
+declare const value: TestEnum;
+declare const valueOrNull: TestEnum | null;
+declare const valueOrUndefined: TestEnum | undefined;
+
 const enumWrapper = $enum(TestEnum);
 
-// $ExpectType EnumWrapper<string | number, { A: number; B: number; C: string; }>
+// $ExpectType EnumWrapper<string, typeof TestEnum>
 enumWrapper;
 
 // $ExpectType number
@@ -38,7 +38,7 @@ enumWrapper.size;
 // $ExpectError
 enumWrapper.size = 0; // immutable
 
-// $ExpectType Readonly<["A" | "B" | "C", string | number]>
+// $ExpectType Readonly<["A" | "B" | "C", TestEnum]>
 enumWrapper[0];
 // $ExpectError
 enumWrapper[0] = ["A", TestEnum.A]; // immutable
@@ -50,10 +50,10 @@ enumWrapper[0][1] = TestEnum.A; // immutable
 // $ExpectType IterableIterator<"A" | "B" | "C">
 enumWrapper.keys();
 
-// $ExpectType IterableIterator<string | number>
+// $ExpectType IterableIterator<TestEnum>
 enumWrapper.values();
 
-// $ExpectType IterableIterator<Readonly<["A" | "B" | "C", string | number]>>
+// $ExpectType IterableIterator<Readonly<["A" | "B" | "C", TestEnum]>>
 enumWrapper.entries();
 for (const entry of enumWrapper.entries()) {
     // $ExpectError
@@ -64,11 +64,11 @@ for (const entry of enumWrapper.entries()) {
 
 // $ExpectType void
 enumWrapper.forEach((value, key, collection, index) => {
-    // $ExpectType string | number
+    // $ExpectType TestEnum
     value;
     // $ExpectType "A" | "B" | "C"
     key;
-    // $ExpectType EnumWrapper<string | number, { A: number; B: number; C: string; }>
+    // $ExpectType EnumWrapper<string, typeof TestEnum>
     collection;
     // $ExpectType number
     index;
@@ -78,11 +78,11 @@ enumWrapper.forEach((value, key, collection, index) => {
 
 // $ExpectType number[]
 enumWrapper.map((value, key, collection, index) => {
-    // $ExpectType string | number
+    // $ExpectType TestEnum
     value;
     // $ExpectType "A" | "B" | "C"
     key;
-    // $ExpectType EnumWrapper<string | number, { A: number; B: number; C: string; }>
+    // $ExpectType EnumWrapper<string, typeof TestEnum>
     collection;
     // $ExpectType number
     index;
@@ -93,10 +93,10 @@ enumWrapper.map((value, key, collection, index) => {
 // $ExpectType ("A" | "B" | "C")[]
 enumWrapper.getKeys();
 
-// $ExpectType (string | number)[]
+// $ExpectType TestEnum[]
 enumWrapper.getValues();
 
-// $ExpectType Readonly<["A" | "B" | "C", string | number]>[]
+// $ExpectType Readonly<["A" | "B" | "C", TestEnum]>[]
 enumWrapper.getEntries();
 const entry = enumWrapper.getEntries()[0];
 // $ExpectError
@@ -151,102 +151,116 @@ enumWrapper.asKeyOrDefault(str, str);
 enumWrapper.asKeyOrDefault(str, strOrUndefined);
 
 // $ExpectType boolean
-enumWrapper.isValue(numstr);
+enumWrapper.isValue(str);
 // $ExpectType boolean
-enumWrapper.isValue(numstrOrNull);
+enumWrapper.isValue(strOrNull);
 // $ExpectType boolean
-enumWrapper.isValue(numstrOrUndefined);
+enumWrapper.isValue(strOrUndefined);
+// $ExpectError
+enumWrapper.isValue(num);
 
-if (enumWrapper.isValue(numstr)) {
-    // $ExpectType string | number
-    numstr;
+if (enumWrapper.isValue(str)) {
+    // $ExpectType TestEnum
+    str;
 }
 
-if (enumWrapper.isValue(numstrOrNull)) {
-    // $ExpectType string | number
-    numstrOrNull;
+if (enumWrapper.isValue(strOrNull)) {
+    // $ExpectType TestEnum
+    strOrNull;
 }
 
-if (enumWrapper.isValue(numstrOrUndefined)) {
-    // $ExpectType string | number
-    numstrOrUndefined;
+if (enumWrapper.isValue(strOrUndefined)) {
+    // $ExpectType TestEnum
+    strOrUndefined;
 }
 
-// $ExpectType string | number
-enumWrapper.asValueOrThrow(numstr);
-// $ExpectType string | number
-enumWrapper.asValueOrThrow(numstrOrNull);
-// $ExpectType string | number
-enumWrapper.asValueOrThrow(numstrOrUndefined);
+// $ExpectType TestEnum
+enumWrapper.asValueOrThrow(str);
+// $ExpectType TestEnum
+enumWrapper.asValueOrThrow(strOrNull);
+// $ExpectType TestEnum
+enumWrapper.asValueOrThrow(strOrUndefined);
+// $ExpectError
+enumWrapper.asValueOrThrow(num);
 
-// $ExpectType string | number | undefined
-enumWrapper.asValueOrDefault(numstr);
-// $ExpectType string | number | undefined
-enumWrapper.asValueOrDefault(numstrOrNull);
-// $ExpectType string | number | undefined
-enumWrapper.asValueOrDefault(numstrOrUndefined);
+// $ExpectType TestEnum | undefined
+enumWrapper.asValueOrDefault(str);
+// $ExpectType TestEnum | undefined
+enumWrapper.asValueOrDefault(strOrNull);
+// $ExpectType TestEnum | undefined
+enumWrapper.asValueOrDefault(strOrUndefined);
+// $ExpectError
+enumWrapper.asValueOrDefault(num);
 
-// $ExpectType string | number | undefined
-enumWrapper.asValueOrDefault(numstr, undefined);
-// $ExpectType string | number
-enumWrapper.asValueOrDefault(numstr, num);
-// $ExpectType string | number
-enumWrapper.asValueOrDefault(numstr, str);
-// $ExpectType string | number
-enumWrapper.asValueOrDefault(numstr, numstr);
-// $ExpectType string | number | undefined
-enumWrapper.asValueOrDefault(num, numstrOrUndefined);
-
-// $ExpectType "A" | "B" | "C"
-enumWrapper.getKeyOrThrow(numstr);
-// $ExpectType "A" | "B" | "C"
-enumWrapper.getKeyOrThrow(numstrOrNull);
-// $ExpectType "A" | "B" | "C"
-enumWrapper.getKeyOrThrow(numstrOrUndefined);
-
-// $ExpectType "A" | "B" | "C" | undefined
-enumWrapper.getKeyOrDefault(numstr);
-// $ExpectType "A" | "B" | "C" | undefined
-enumWrapper.getKeyOrDefault(numstrOrNull);
-// $ExpectType "A" | "B" | "C" | undefined
-enumWrapper.getKeyOrDefault(numstrOrUndefined);
-
-// $ExpectType "A" | "B" | "C"
-enumWrapper.getKeyOrDefault(numstr, key);
-// $ExpectType "A" | "B" | "C" | undefined
-enumWrapper.getKeyOrDefault(numstr, keyOrUndefined);
+// $ExpectType TestEnum | undefined
+enumWrapper.asValueOrDefault(str, undefined);
+// $ExpectType TestEnum
+enumWrapper.asValueOrDefault(str, value);
+// $ExpectType TestEnum | undefined
+enumWrapper.asValueOrDefault(str, valueOrUndefined);
 // $ExpectType string
-enumWrapper.getKeyOrDefault(numstr, str);
+enumWrapper.asValueOrDefault(str, str);
 // $ExpectType string | undefined
-enumWrapper.getKeyOrDefault(numstr, strOrUndefined);
+enumWrapper.asValueOrDefault(str, strOrUndefined);
+// $ExpectError
+enumWrapper.asValueOrDefault(str, num);
 
-// $ExpectType string | number
+// $ExpectType "A" | "B" | "C"
+enumWrapper.getKeyOrThrow(str);
+// $ExpectType "A" | "B" | "C"
+enumWrapper.getKeyOrThrow(strOrNull);
+// $ExpectType "A" | "B" | "C"
+enumWrapper.getKeyOrThrow(strOrUndefined);
+// $ExpectError
+enumWrapper.getKeyOrThrow(num);
+
+// $ExpectType "A" | "B" | "C" | undefined
+enumWrapper.getKeyOrDefault(str);
+// $ExpectType "A" | "B" | "C" | undefined
+enumWrapper.getKeyOrDefault(strOrNull);
+// $ExpectType "A" | "B" | "C" | undefined
+enumWrapper.getKeyOrDefault(strOrUndefined);
+// $ExpectError
+enumWrapper.getKeyOrDefault(num);
+
+// $ExpectType "A" | "B" | "C"
+enumWrapper.getKeyOrDefault(str, key);
+// $ExpectType "A" | "B" | "C" | undefined
+enumWrapper.getKeyOrDefault(str, keyOrUndefined);
+// $ExpectType string
+enumWrapper.getKeyOrDefault(str, str);
+// $ExpectType string | undefined
+enumWrapper.getKeyOrDefault(str, strOrUndefined);
+
+// $ExpectType TestEnum
 enumWrapper.getValueOrThrow(key);
-// $ExpectType string | number
+// $ExpectType TestEnum
 enumWrapper.getValueOrThrow(keyOrNull);
-// $ExpectType string | number
+// $ExpectType TestEnum
 enumWrapper.getValueOrThrow(keyOrUndefined);
-// $ExpectType string | number
+// $ExpectType TestEnum
 enumWrapper.getValueOrThrow(str);
-// $ExpectType string | number
+// $ExpectType TestEnum
 enumWrapper.getValueOrThrow(strOrNull);
-// $ExpectType string | number
+// $ExpectType TestEnum
 enumWrapper.getValueOrThrow(strOrUndefined);
 
-// $ExpectType string | number | undefined
+// $ExpectType TestEnum | undefined
 enumWrapper.getValueOrDefault(str);
-// $ExpectType string | number | undefined
+// $ExpectType TestEnum | undefined
 enumWrapper.getValueOrDefault(strOrNull);
-// $ExpectType string | number | undefined
+// $ExpectType TestEnum | undefined
 enumWrapper.getValueOrDefault(strOrUndefined);
 
-// $ExpectType string | number | undefined
+// $ExpectType TestEnum | undefined
 enumWrapper.getValueOrDefault(str, undefined);
-// $ExpectType string | number
-enumWrapper.getValueOrDefault(str, num);
-// $ExpectType string | number
+// $ExpectType TestEnum
+enumWrapper.getValueOrDefault(str, value);
+// $ExpectType TestEnum | undefined
+enumWrapper.getValueOrDefault(str, valueOrUndefined);
+// $ExpectType string
 enumWrapper.getValueOrDefault(str, str);
-// $ExpectType string | number
-enumWrapper.getValueOrDefault(str, numstr);
-// $ExpectType string | number | undefined
-enumWrapper.getValueOrDefault(str, numstrOrUndefined);
+// $ExpectType string | undefined
+enumWrapper.getValueOrDefault(str, strOrUndefined);
+// $ExpectError
+enumWrapper.getValueOrDefault(str, num);
